requirements.txt
requests 
websocket-client
psutil
OpenAI